+++ packages/x-charts/src/ChartsAxisHighlight/ChartsYAxisHighlight.tsx
@@ -4,7 +4,7 @@ import { getValueToPositionMapper, useYScale } from '../hooks/useScale';
 import { isBandScale } from '../internals/isBandScale';
 import { useSelector } from '../internals/store/useSelector';
 import { useStore } from '../internals/store/useStore';
-import { selectorChartsInteractionYAxis } from '../context/InteractionSelectors';
+import { selectorChartsInteractionYAxis } from '../internals/plugins/featurePlugins/useChartInteraction';
 import { useDrawingArea } from '../hooks';
 import { ChartsAxisHighlightType } from './ChartsAxisHighlight.types';
 import { ChartsAxisHighlightClasses } from './chartsAxisHighlightClasses';
