+++ packages/x-charts/src/ScatterChart/Scatter.tsx
@@ -13,7 +13,7 @@ import { useSelector } from '../internals/store/useSelector';
 import { D3Scale } from '../models/axis';
 import { useHighlighted } from '../context';
 import { useDrawingArea } from '../hooks/useDrawingArea';
-import { selectorChartsInteractionIsVoronoiEnabled } from '../context/InteractionSelectors';
+import { selectorChartsInteractionIsVoronoiEnabled } from '../internals/plugins/featurePlugins/useChartInteraction';
 
 export interface ScatterProps {
   series: DefaultizedScatterSeriesType;
