+++ packages/x-charts/src/internals/plugins/featurePlugins/useChartInteraction/index.ts
@@ -1,2 +1,3 @@
 export { useChartInteraction } from './useChartInteraction';
+export * from './useChartInteraction.selectors';
 export type { UseChartInteractionSignature } from './useChartInteraction.types';
